Revert "184 documentation its time to get working"
