+++ src/main/java/frc/robot/Robot.java
@@ -13,10 +13,10 @@ import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
-import frc.robot.util.constants.Constants.AutoConstants;
-import frc.robot.util.constants.Constants.DriveConstants;
-import frc.robot.util.constants.Constants.NeoMotorConstants;
-import frc.robot.util.motors.Neo;
+import frc.robot.util.Neo;
+import frc.robot.util.Constants.AutoConstants;
+import frc.robot.util.Constants.DriveConstants;
+import frc.robot.util.Constants.NeoMotorConstants;
 import monologue.Monologue;
