+++ src/main/java/frc/robot/commands/Drive.java
@@ -1,13 +1,15 @@
-package frc.robot.commands.drive;
+package frc.robot.commands;
 
 import java.util.function.BooleanSupplier;
 import java.util.function.DoubleSupplier;
 import java.util.function.Supplier;
 
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
+//import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.Swerve;
-import frc.robot.util.constants.Constants.DriveConstants;
+//import frc.robot.util.Constants.FieldConstants;
+import frc.robot.util.Constants.DriveConstants;
 
 public class Drive extends Command {
