+++ src/main/java/frc/robot/commands/DriveHDC.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.drive;
+package frc.robot.commands;
 
 import java.util.function.BooleanSupplier;
 import java.util.function.DoubleSupplier;
@@ -10,9 +10,9 @@ import edu.wpi.first.math.kinematics.ChassisSpeeds;
 //import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.Swerve;
-import frc.robot.util.constants.Constants.AutoConstants;
-import frc.robot.util.constants.Constants.DriveConstants;
-import frc.robot.util.testing.HDCTuner;
+import frc.robot.util.HDCTuner;
+import frc.robot.util.Constants.AutoConstants;
+import frc.robot.util.Constants.DriveConstants;
 
 public class DriveHDC  extends Command {
