+++ src/main/java/frc/robot/commands/autonomous/ChoreoStorage.java
@@ -8,8 +8,8 @@ import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.util.constants.Constants.AutoConstants;
-import frc.robot.util.constants.Constants.FieldConstants;
+import frc.robot.util.Constants.AutoConstants;
+import frc.robot.util.Constants.FieldConstants;
 import monologue.Logged;
 import monologue.Annotations.Log;
