+++ src/main/java/frc/robot/subsystems/Intake.java
@@ -5,9 +5,9 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.Robot;
-import frc.robot.util.constants.Constants.IntakeConstants;
-import frc.robot.util.motors.Neo;
-import frc.robot.util.motors.SafeSparkMax.TelemetryPreference;
+import frc.robot.util.Neo;
+import frc.robot.util.Constants.IntakeConstants;
+import frc.robot.util.SafeSpark.TelemetryPreference;
 import monologue.Logged;
 import monologue.Annotations.Log;
