+++ src/main/java/frc/robot/subsystems/Swerve.java
@@ -30,14 +30,14 @@ import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Robot;
 import frc.robot.RobotContainer;
-import frc.robot.commands.drive.Drive;
-import frc.robot.commands.drive.DriveHDC;
-import frc.robot.util.constants.Constants.AutoConstants;
-import frc.robot.util.constants.Constants.DriveConstants;
-import frc.robot.util.constants.Constants.FieldConstants;
-import frc.robot.util.constants.Constants.ModuleConstants;
-import frc.robot.util.motors.MAXSwerveModule;
-import frc.robot.util.testing.PIDNotConstants;
+import frc.robot.commands.Drive;
+import frc.robot.commands.DriveHDC;
+import frc.robot.util.MAXSwerveModule;
+import frc.robot.util.PIDNotConstants;
+import frc.robot.util.Constants.AutoConstants;
+import frc.robot.util.Constants.DriveConstants;
+import frc.robot.util.Constants.FieldConstants;
+import frc.robot.util.Constants.ModuleConstants;
 import monologue.Logged;
 import monologue.Annotations.Log;
