+++ src/main/java/frc/robot/subsystems/elevator/Elevator.java
@@ -7,10 +7,11 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.RobotContainer;
-import frc.robot.util.constants.Constants.NTConstants;
-import frc.robot.util.constants.Constants.TrapConstants;
-import frc.robot.util.motors.Neo;
-import frc.robot.util.testing.PIDNotConstants;
+import frc.robot.util.Neo;
+import frc.robot.util.PIDNotConstants;
+import frc.robot.util.Constants.NTConstants;
+import frc.robot.util.Constants.TrapConstants;
+import frc.robot.util.SafeSpark.TelemetryPreference;
 import monologue.Logged;
 import monologue.Annotations.Log;
