+++ src/main/java/frc/robot/subsystems/shooter/Shooter.java
@@ -5,9 +5,10 @@ import edu.wpi.first.math.Pair;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.util.constants.Constants.ShooterConstants;
-import frc.robot.util.motors.Neo;
-import frc.robot.util.testing.PIDNotConstants;
+import frc.robot.util.Neo;
+import frc.robot.util.Constants.ShooterConstants;
+import frc.robot.util.SafeSpark.TelemetryPreference;
+import frc.robot.util.PIDNotConstants;
 import monologue.Logged;
 import monologue.Annotations.Log;
