+++ src/main/java/frc/robot/util/HDCTuner.java
@@ -1,4 +1,4 @@
-package frc.robot.util.testing;
+package frc.robot.util;
 
 import edu.wpi.first.math.controller.HolonomicDriveController;
 import edu.wpi.first.math.controller.PIDController;
@@ -6,7 +6,7 @@ import edu.wpi.first.math.controller.ProfiledPIDController;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.util.constants.Constants.AutoConstants;
+import frc.robot.util.Constants.AutoConstants;
 import monologue.Annotations.Log;
 import monologue.Logged;
