+++ src/main/java/frc/robot/util/LimelightHelpers.java
@@ -1,6 +1,6 @@
 //LimelightHelpers v1.2.1 (March 1, 2023)
 // https://github.com/LimelightVision/limelightlib-wpijava/
-package frc.robot.util.calc;
+package frc.robot.util;
 
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableEntry;
