+++ src/main/java/frc/robot/util/MAXSwerveModule.java
@@ -2,16 +2,15 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.util.motors;
+package frc.robot.util;
 
 import com.revrobotics.SparkPIDController;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
-import frc.robot.util.constants.Constants.FieldConstants;
-import frc.robot.util.constants.Constants.ModuleConstants;
-import frc.robot.util.testing.PIDNotConstants;
+import frc.robot.util.Constants.FieldConstants;
+import frc.robot.util.Constants.ModuleConstants;
 import monologue.Logged;
 import monologue.Annotations.Log;
