+++ src/main/java/frc/robot/util/PatriSendableChooser.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.util.mod;
+package frc.robot.util;
 
 import static edu.wpi.first.util.ErrorMessages.requireNonNullParam;
 
@@ -156,43 +156,43 @@ public class PatriSendableChooser<V> implements Sendable, AutoCloseable {
         builder.addStringProperty(DEFAULT, () -> m_defaultChoice, null);
         builder.addStringArrayProperty(OPTIONS, () -> m_map.keySet().toArray(new String[0]), null);
         builder.addStringProperty(
-            ACTIVE,
-            () -> {
-                m_mutex.lock();
-                try {
-                    if (m_selected != null) {
-                        return m_selected;
-                    } else {
-                        return m_defaultChoice;
+                ACTIVE,
+                () -> {
+                    m_mutex.lock();
+                    try {
+                        if (m_selected != null) {
+                            return m_selected;
+                        } else {
+                            return m_defaultChoice;
+                        }
+                    } finally {
+                        m_mutex.unlock();
                     }
-                } finally {
-                    m_mutex.unlock();
-                }
-            },
-            null);
+                },
+                null);
         builder.addStringProperty(
-            SELECTED,
-            null,
-            val -> {
-                V choice;
-                Consumer<V> listener;
-                m_mutex.lock();
-                try {
-                    m_selected = val;
-                    if (!m_selected.equals(m_previousVal) && m_listener != null) {
-                        choice = m_map.get(val);
-                        listener = m_listener;
-                    } else {
-                        choice = null;
-                        listener = null;
+                SELECTED,
+                null,
+                val -> {
+                    V choice;
+                    Consumer<V> listener;
+                    m_mutex.lock();
+                    try {
+                        m_selected = val;
+                        if (!m_selected.equals(m_previousVal) && m_listener != null) {
+                            choice = m_map.get(val);
+                            listener = m_listener;
+                        } else {
+                            choice = null;
+                            listener = null;
+                        }
+                        m_previousVal = val;
+                    } finally {
+                        m_mutex.unlock();
                     }
-                    m_previousVal = val;
-                } finally {
-                    m_mutex.unlock();
-                }
-                if (listener != null) {
-                    listener.accept(choice);
-                }
-            });
+                    if (listener != null) {
+                        listener.accept(choice);
+                    }
+                });
     }
 }
